feat: make the package proper template
This commit makes the package a proper template by adding:
 - name and version (required by npm)
 - templateVersion - to indicate to NativeScript CLI the proper way to handle the files during project creation
 - updating dependencies and runtimes to 6.4 version. NOTE: The major and minor version of the template should match the major and minor version of the other NativeScript packages (for example tns-core-modules).

Also use the new feature introduced in NativeScript CLI 6.4 for using custom webpack configuration file. This path to file is set in the nsconfig.json and all the custom logic can be moved in it. The `webpack.config.js` file is removed, gitignored and npmignored to ensure the users will receive its latest version from the postinstall script of nativescript-dev-webpack.